Move Client and Auth into lib/sendgrid4r/
