+++ lib/sendgrid4r.rb
@@ -1,5 +1,5 @@
 # -*- encoding: utf-8 -*-
 $LOAD_PATH.unshift File.dirname(__FILE__)
 
-require 'client'
+require 'sendgrid4r/client'
 require 'sendgrid4r/version'
