+++ lib/sendgrid4r/auth.rb
