+++ lib/sendgrid4r/client.rb
@@ -1,7 +1,7 @@
 # -*- encoding: utf-8 -*-
 $LOAD_PATH.unshift File.dirname(__FILE__)
 
-require 'auth'
+require 'sendgrid4r/auth'
 require 'sendgrid4r/rest/api'
 require 'sendgrid4r/factory/version_factory'
 require 'sendgrid4r/factory/condition_factory'
